+++ src/main/java/com/example/hotel/blImpl/coupon/TargetMoneyCouponStrategy.java
@@ -2,8 +2,8 @@ package com.example.hotel.blImpl.coupon;
 
 import com.example.hotel.bl.coupon.CouponMatchStrategyI;
 import com.example.hotel.enums.CouponType;
-import com.example.hotel.po.Coupon;
-import com.example.hotel.po.Order;
+import com.example.hotel.model.Coupon;
+import com.example.hotel.model.Order;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -12,7 +12,7 @@ public class TargetMoneyCouponStrategy implements CouponMatchStrategyI {
     @Override
     public boolean isMatch(Order order, Coupon coupon) {
         if (coupon.getType() != CouponType.Target) return false;
-        return order.getPrice() >= coupon.getTargetMoney();
+        return order.getPrice().doubleValue() >= coupon.getTargetMoney();
     }
 
 }
